Update to latest CN1 source

+++ Ports/iOSPort/nativeSources/OpenUDID.m
@@ -135,7 +135,7 @@ static int const kOpenUDIDRedundancySlots = 100;
         CFStringRef cfstring = CFUUIDCreateString(kCFAllocatorDefault, uuid);
         const char *cStr = CFStringGetCStringPtr(cfstring,CFStringGetFastestEncoding(cfstring));
         unsigned char result[16];
-        CC_MD5( cStr, strlen(cStr), result );
+        CC_MD5( cStr, (unsigned int)strlen(cStr), result );
         CFRelease(uuid);
 
         _openUDID = [NSString stringWithFormat:
@@ -144,7 +144,7 @@ static int const kOpenUDIDRedundancySlots = 100;
                 result[4], result[5], result[6], result[7],
                 result[8], result[9], result[10], result[11],
                 result[12], result[13], result[14], result[15],
-                     (NSUInteger)(arc4random() % NSUIntegerMax)];  
+                     (unsigned long)(arc4random() % NSUIntegerMax)];  
     }
     
     // Call to other developers in the Open Source community:
